refactor: 깃허브 api route, controller, util 분리
