+++ backend/src/server.ts
@@ -7,8 +7,9 @@ dotenv.config();
 
 const app = express();
 const PORT = process.env.PORT || 5000;
+const CLIENT_URL = process.env.CLIENT_URL;
 
-app.use(cors({ origin: "http://localhost:5173" }));
+app.use(cors({ origin: CLIENT_URL }));
 app.use(express.json());
 
 // Routes
