+++ docs/checklist.md
@@ -71,6 +71,7 @@
 - [ ] cva 사용하여 컴포넌트 조건부 스타일링 (Button 등)
 - [ ] Icon 컴포넌트 구현
 - [ ] 프로젝트 구조 개선
+- [ ] 서버 라우트와 컨트롤러 분리
 
 ---
